requirements.txt
plotly==5.20.0
gspread==6.1.0
streamlit-authenticator
python-telegram-bot